add : Validations in BaseAnimal

+++ 13. Exams practice/Java OOP Retake Exam - 18 April 2022/01. High Quality Structure + 02. Business Logic/ZooApp/src/main/java/validator/Validator.java
@@ -22,7 +22,7 @@ public class Validator {
     }
   }
 
-  public static void validateAnimalPrice(int price) {
+  public static void validateAnimalPrice(double price) {
     if (price <= 0) {
       throw new IllegalArgumentException(ExceptionMessages.ANIMAL_PRICE_BELOW_OR_EQUAL_ZERO);
     }
